+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# 0.0.1
+## 0.1.0
+
+* Add a new parser - `cherow`.
+* Add tests including the source code locations in the AST.
+* Add tests of code generators.
+
+## 0.0.1
 
 Initial release showing script and module parsing speeds and package sizes.